health_app/requirements.txt
Flask==2.3.2
requests==2.31.0
openai==0.28
python-dotenv==1.0.0
beautifulsoup4==4.12.2
